Fix microsoft-contrib style guide
