+++ .editorconfig
@@ -6,3 +6,4 @@ charset = utf-8
 indent_style = space
 indent_size = 4
 trim_trailing_whitespace = true
+insert_final_newline = true
