--- test/data.ts
@@ -1,303 +1,0 @@
-/*
- *  Power BI Visualizations
- *
- *  Copyright (c) Microsoft Corporation
- *  All rights reserved.
- *  MIT License
- *
- *  Permission is hereby granted, free of charge, to any person obtaining a copy
- *  of this software and associated documentation files (the ""Software""), to deal
- *  in the Software without restriction, including without limitation the rights
- *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- *  copies of the Software, and to permit persons to whom the Software is
- *  furnished to do so, subject to the following conditions:
- *
- *  The above copyright notice and this permission notice shall be included in
- *  all copies or substantial portions of the Software.
- *
- *  THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- *  THE SOFTWARE.
- */
-
-/// <reference path="_references.ts"/>
-
-module powerbi.extensibility.visual.test.data {
-  export const Countries = [
-    { name: "Afghanistan", code: "AF" },
-    { name: "Åland Islands", code: "AX" },
-    { name: "Albania", code: "AL" },
-    { name: "Algeria", code: "DZ" },
-    { name: "American Samoa", code: "AS" },
-    { name: "AndorrA", code: "AD" },
-    { name: "Angola", code: "AO" },
-    { name: "Anguilla", code: "AI" },
-    { name: "Antarctica", code: "AQ" },
-    { name: "Antigua and Barbuda", code: "AG" },
-    { name: "Argentina", code: "AR" },
-    { name: "Armenia", code: "AM" },
-    { name: "Aruba", code: "AW" },
-    { name: "Australia", code: "AU" },
-    { name: "Austria", code: "AT" },
-    { name: "Azerbaijan", code: "AZ" },
-    { name: "Bahamas", code: "BS" },
-    { name: "Bahrain", code: "BH" },
-    { name: "Bangladesh", code: "BD" },
-    { name: "Barbados", code: "BB" },
-    { name: "Belarus", code: "BY" },
-    { name: "Belgium", code: "BE" },
-    { name: "Belize", code: "BZ" },
-    { name: "Benin", code: "BJ" },
-    { name: "Bermuda", code: "BM" },
-    { name: "Bhutan", code: "BT" },
-    { name: "Bolivia", code: "BO" },
-    { name: "Bosnia and Herzegovina", code: "BA" },
-    { name: "Botswana", code: "BW" },
-    { name: "Bouvet Island", code: "BV" },
-    { name: "Brazil", code: "BR" },
-    { name: "British Indian Ocean Territory", code: "IO" },
-    { name: "Brunei Darussalam", code: "BN" },
-    { name: "Bulgaria", code: "BG" },
-    { name: "Burkina Faso", code: "BF" },
-    { name: "Burundi", code: "BI" },
-    { name: "Cambodia", code: "KH" },
-    { name: "Cameroon", code: "CM" },
-    { name: "Canada", code: "CA" },
-    { name: "Cape Verde", code: "CV" },
-    { name: "Cayman Islands", code: "KY" },
-    { name: "Central African Republic", code: "CF" },
-    { name: "Chad", code: "TD" },
-    { name: "Chile", code: "CL" },
-    { name: "China", code: "CN" },
-    { name: "Christmas Island", code: "CX" },
-    { name: "Cocos (Keeling) Islands", code: "CC" },
-    { name: "Colombia", code: "CO" },
-    { name: "Comoros", code: "KM" },
-    { name: "Congo", code: "CG" },
-    { name: "Congo, The Democratic Republic of the", code: "CD" },
-    { name: "Cook Islands", code: "CK" },
-    { name: "Costa Rica", code: "CR" },
-    { name: "Cote D\"Ivoire", code: "CI" },
-    { name: "Croatia", code: "HR" },
-    { name: "Cuba", code: "CU" },
-    { name: "Cyprus", code: "CY" },
-    { name: "Czech Republic", code: "CZ" },
-  ];
-
-  export const States = [
-    {
-      name: "ALABAMA",
-      code: "AL"
-    },
-    {
-      name: "ALASKA",
-      code: "AK"
-    },
-    {
-      name: "ARIZONA",
-      code: "AZ"
-    },
-    {
-      name: "ARKANSAS",
-      code: "AR"
-    },
-    {
-      name: "CALIFORNIA",
-      code: "CA"
-    },
-    {
-      name: "COLORADO",
-      code: "CO"
-    },
-    {
-      name: "CONNECTICUT",
-      code: "CT"
-    },
-    {
-      name: "DELAWARE",
-      code: "DE"
-    },
-    {
-      name: "FLORIDA",
-      code: "FL"
-    },
-    {
-      name: "GEORGIA",
-      code: "GA"
-    },
-    {
-      name: "HAWAII",
-      code: "HI"
-    },
-    {
-      name: "IDAHO",
-      code: "ID"
-    },
-    {
-      name: "ILLINOIS",
-      code: "IL"
-    },
-    {
-      name: "INDIANA",
-      code: "IN"
-    },
-    {
-      name: "IOWA",
-      code: "IA"
-    },
-    {
-      name: "KANSAS",
-      code: "KS"
-    },
-    {
-      name: "KENTUCKY",
-      code: "KY"
-    },
-    {
-      name: "LOUISIANA",
-      code: "LA"
-    },
-    {
-      name: "MAINE",
-      code: "ME"
-    },
-    {
-      name: "MARYLAND",
-      code: "MD"
-    },
-    {
-      name: "MASSACHUSETTS",
-      code: "MA"
-    },
-    {
-      name: "MICHIGAN",
-      code: "MI"
-    },
-    {
-      name: "MINNESOTA",
-      code: "MN"
-    },
-    {
-      name: "MISSISSIPPI",
-      code: "MS"
-    },
-    {
-      name: "MISSOURI",
-      code: "MO"
-    },
-    {
-      name: "MONTANA",
-      code: "MT"
-    },
-    {
-      name: "NEBRASKA",
-      code: "NE"
-    },
-    {
-      name: "NEVADA",
-      code: "NV"
-    },
-    {
-      name: "NEW HAMPSHIRE",
-      code: "NH"
-    },
-    {
-      name: "NEW JERSEY",
-      code: "NJ"
-    },
-    {
-      name: "NEW MEXICO",
-      code: "NM"
-    },
-    {
-      name: "NEW YORK",
-      code: "NY"
-    },
-    {
-      name: "NORTH CAROLINA",
-      code: "NC"
-    },
-    {
-      name: "NORTH DAKOTA",
-      code: "ND"
-    },
-    {
-      name: "OHIO",
-      code: "OH"
-    },
-    {
-      name: "OKLAHOMA",
-      code: "OK"
-    },
-    {
-      name: "OREGON",
-      code: "OR"
-    },
-    {
-      name: "PENNSYLVANIA",
-      code: "PA"
-    },
-    {
-      name: "RHODE ISLAND",
-      code: "RI"
-    },
-    {
-      name: "SOUTH CAROLINA",
-      code: "SC"
-    },
-    {
-      name: "SOUTH DAKOTA",
-      code: "SD"
-    },
-    {
-      name: "TENNESSEE",
-      code: "TN"
-    },
-    {
-      name: "TEXAS",
-      code: "TX"
-    },
-    {
-      name: "UTAH",
-      code: "UT"
-    },
-    {
-      name: "VERMONT",
-      code: "VT"
-    },
-    {
-      name: "VIRGINIA",
-      code: "VA"
-    },
-    {
-      name: "WASHINGTON",
-      code: "WA"
-    },
-    {
-      name: "WEST VIRGINIA",
-      code: "WV"
-    },
-    {
-      name: "WISCONSIN",
-      code: "WI"
-    },
-    {
-      name: "WYOMING",
-      code: "WY"
-    }
-  ];
-
-  export const Regions = [
-    { name: "Asia" },
-    { name: "Europe" },
-    { name: "Africa" },
-    { name: "Oceania" },
-    { name: "North America" },
-    { name: "Antarctica" },
-    { name: "South America" }
-  ];
-}
