+++ tsconfig.json
@@ -8,21 +8,16 @@
     "sourceMap": true,
     "sourceRoot": "../../src/",
     "out": "./.tmp/build/visual.js",
-    "typeRoots": [
-      "node_modules/@types",
-      ".api/v1.5.0"
-    ],
     "declaration": true
   },
   "files": [
-    ".api/v1.5.0/PowerBI-visuals.d.ts",
+    ".api/v1.6.0/PowerBI-visuals.d.ts",
     "node_modules/powerbi-visuals-utils-dataviewutils/lib/index.d.ts",
     "node_modules/powerbi-visuals-utils-tooltiputils/lib/index.d.ts",
     "node_modules/powerbi-visuals-utils-typeutils/lib/index.d.ts",
     "node_modules/powerbi-visuals-utils-svgutils/lib/index.d.ts",
-    "node_modules/powerbi-visuals-utils-formattingutils/lib/index.d.ts",
     "src/dataInterfaces.ts",
-    "src/settings.ts",
-    "src/visual.ts"
+    "src/SunburstSettings.ts",
+    "src/SunburstVisual.ts"
   ]
 }
